dev: DropLinks components moved to common folder
