+++ src/components/Header/Header.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import './Header.css';
-import DropLinks from '../DropLinks';
+import DropLinks from '../DropLinks/DropLinks.js';
 import logo from '../../images/turbina-logo.svg';
 import Player from '../Player/Player';
