+++ src/index.css
@@ -2,7 +2,7 @@
 
 @import url('./vendors/fonts.css');
 @import url('./vendors/normalize.css');
-@import url('./components/DropLinks.css');
-@import url('./components/LinkTab.css');
-@import url('./components/LinkHead.css');
-@import url('./components/LinkCloseButton.css');
+@import url('./components/DropLinks/DropLinks.css');
+@import url('./components/DropLinks/LinkTab.css');
+@import url('./components/DropLinks/LinkHead.css');
+@import url('./components/DropLinks/LinkCloseButton.css');
